Move test to test_cuda
